+++ resources/views/pages/admin/knowledge/index.blade.php
@@ -1,0 +1,5 @@
+<x-layouts::app :title="__('Knowledge Admin')">
+    <section class="w-full">
+        <livewire:admin.knowledge.article-index />
+    </section>
+</x-layouts::app>